Patch 0537
・メモリマネージャに関する修正（malloc.c）
	-> ポインタのアドレスを計算するときにはuintptr型を使うように修正
	-> _WIN64のときの制限を撤廃
	-> メモリリーク記録用のログファイル名の生成処理において、
	   バッファオーバーフローが発生する可能性を修正
・numdb_cmp() における大小比較は、ポインタのアドレスでも計算できるように
  intptr型で演算する（db.c）
・不要なキャストを削除（login.c, clif.c, map.h）
・db_init() で、不要な初期化を省略（db.c）

+++ src/login/login.c
@@ -2286,7 +2286,7 @@ void login_socket_ctrl_panel_func(int fd,char* usage,char* user,char* status)
 	}
 	else if( sd->func_parse == parse_login && sd->auth )
 	{
-		sprintf( user, "%s (%d)", ld->userid, (int)ld->account_id );
+		sprintf( user, "%s (%d)", ld->userid, ld->account_id );
 	}
 }
